added dqn with ptan

+++ atari/withPTAN/lib/dqn_model.py
@@ -32,5 +32,5 @@ class DQN(nn.Module):
         return int(np.prod(o.size()))
 
     def forward(self, x):
-        conv_out = self.conv(x).view(x.size()[0], -1)
+        conv_out = self.conv(x.float()).view(x.size()[0], -1)
         return self.fc(conv_out)
